requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==1.10.15
boto3==1.34.144
jira==3.5.2
openai==1.41.0
structlog==24.2.0
tenacity==8.3.0
requests==2.31.0


